Update sketch_orgmode.cpp - added brute force protection
also determined that, yes, XSS via the message prompts works great (cool, but bad: try it like this: <script>alert('You have been hacked');</script>, just put that block in the message, send it, and on refresh the browser will pop up a scary message. Not great, way too easy. We'll have to do something about that. 

The brute force lockout will be a good step. 

We also need to consider the implications of the password being sent in plain text over http. 

Maybe we can do a type of two factor with the TFT screen? It'll prompt if you're sending the message and we change the cred to MAC address? And organizers will have a TFT?