Update sketch_primary - make public enable/disable more robust
adopts status from configured node that sent it OTA password. If it's receiving a password, it's assumed the provisioning device is provisioned. Defaults to false. Better to not allow it if intended to be available than be available if it's considered risky. 

Organizers should set the public=on initially if they choose to. This is intended to reconfigure temporarily disconnected nodes.