Update sketch_primary - lock public off on disable
The logic here is if you want it on, turn it on at the start of the protest. If you want it off, we need modes to honor that no matter what. If it's become dangerous or being vandalized, organizers need to be able to end it quickly and reliably. 
The public message feature is a double edged sword. If there's any doubt if it should be on or off, I want it off.